Add events, payments, tasks, and automation templates
