+++ al-ghadeer-events/docs/USER_MANUAL.md
